requirements.txt
Django==1.4.3
django-registration==0.8
psycopg2==2.4.6
django-autoslug==1.6.1
